move healthcheck router to healthcheck dir
